+++ pkg/apis/networking/v1alpha1/route_spec_options_upgrade.go
@@ -1,0 +1,48 @@
+//
+// DISCLAIMER
+//
+// Copyright 2024 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package v1alpha1
+
+import shared "github.com/arangodb/kube-arangodb/pkg/apis/shared"
+
+type ArangoRouteSpecOptionsUpgrade []ArangoRouteSpecOptionUpgrade
+
+func (a ArangoRouteSpecOptionsUpgrade) Validate() error {
+	return shared.ValidateInterfaceList(a)
+}
+
+type ArangoRouteSpecOptionUpgrade struct {
+	// Type defines type of the Upgrade
+	// +doc/enum: websocket|HTTP WebSocket Upgrade type
+	Type ArangoRouteUpgradeOptionType `json:"type"`
+
+	// Enabled defines if upgrade option is enabled
+	Enabled *bool `json:"enabled,omitempty"`
+}
+
+func (a ArangoRouteSpecOptionUpgrade) Validate() error {
+	if err := shared.WithErrors(
+		shared.ValidateRequiredInterfacePath("type", a.Type),
+	); err != nil {
+		return err
+	}
+
+	return nil
+}